refactor classes with #booted module.
