+++ apprentice_teacher.rb
@@ -1,6 +1,10 @@
+require_relative 'booted'
+
 class ApprenticeTeacher
   attr_reader :age, :salary, :phase, :target_raise
   attr_accessor :name
+
+  include Booted
 
   def initialize(options={})
     @age = options.fetch(:age, 0)
@@ -9,14 +13,14 @@ class ApprenticeTeacher
     @phase = 3
   end
 
-  def offer_high_five
-    "High five!"
-  end
-
-  def set_phase(num)
-    @phase = num
-    "Cool, I've always wanted to teach phase #{num}!"
-  end
+  # def offer_high_five
+  #   "High five!"
+  # end
+  #
+  # def set_phase(num)
+  #   @phase = num
+  #   "Cool, I've always wanted to teach phase #{num}!"
+  # end
 
   def teach_stuff
     response = ""
