+++ booted.rb
@@ -1,0 +1,22 @@
+module Booted
+
+  def offer_high_five
+    "High five!"
+  end
+
+  def set_phase(num)
+    response = ""
+    if self.class == Student
+      if num == @phase
+        response += "I'm doing phase #{@phase} again because "
+        response += "I put my learning first. I'm gonna rock it!"
+      else
+        response = "Oooh, phase #{num}. I hope I'm ready!"
+      end
+    else
+      response = "Cool, I've always wanted to teach phase #{num}!"
+    end
+    @phase = num
+    response
+  end
+end
