+++ student.rb
@@ -1,6 +1,10 @@
+require_relative 'booted'
+
 class Student
   attr_reader :age, :phase
   attr_accessor :name
+
+  include Booted
 
   def initialize(options = {})
     @phase = 1
@@ -8,21 +12,21 @@ class Student
     @name = options.fetch(:name, "")
   end
 
-  def offer_high_five
-    "High five!"
-  end
-
-  def set_phase(num)
-    response = ""
-    if num == @phase
-      response += "I'm doing phase #{@phase} again because "
-      response += "I put my learning first. I'm gonna rock it!"
-    else
-      response = "Oooh, phase #{num}. I hope I'm ready!"
-    end
-    @phase = num
-    response
-  end
+  # def offer_high_five
+  #   "High five!"
+  # end
+  #
+  # def set_phase(num)
+  #   response = ""
+  #   if num == @phase
+  #     response += "I'm doing phase #{@phase} again because "
+  #     response += "I put my learning first. I'm gonna rock it!"
+  #   else
+  #     response = "Oooh, phase #{num}. I hope I'm ready!"
+  #   end
+  #   @phase = num
+  #   response
+  # end
 
   def learn_stuff
     response = ""
